tests/requirements.txt
pytest>=7.4.0
hypothesis>=6.92.0
pyyaml>=6.0
pytest-html>=3.1.0
cfn-lint>=0.83.0
